+++ src/app/containers/MostRead/Canonical/index.test.jsx
@@ -1,10 +1,3 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import { act } from 'react-dom/test-utils';
-import MostReadContainer from '..';
-import { RequestContextProvider } from '#contexts/RequestContext';
-import { ServiceContextProvider } from '#contexts/ServiceContext';
-import { ToggleContext } from '#contexts/ToggleContext';
 import newsMostReadData from '#data/news/mostRead';
 import zhongwenSimpMostReadData from '#data/zhongwen/mostRead/simp';
 import { service as newsConfig } from '#app/lib/config/services/news';
@@ -12,6 +5,7 @@ import { service as zhongwenConfig } from '#app/lib/config/services/zhongwen';
 import {
   setStalePromoTimestamp,
   setFreshPromoTimestamp,
+  renderMostReadContainer,
 } from '../utilities/testHelper';
 
 let container;
@@ -31,40 +25,6 @@ const services = {
   },
 };
 
-const getToggleState = enabled => ({
-  local: { mostRead: { enabled } },
-  test: { mostRead: { enabled } },
-});
-
-const renderMostReadContainer = async ({
-  service,
-  variant = null,
-  mostReadToggle = false,
-}) =>
-  act(async () => {
-    ReactDOM.render(
-      <ToggleContext.Provider
-        value={{ toggleState: getToggleState(mostReadToggle) }}
-      >
-        <RequestContextProvider
-          bbcOrigin={`http://localhost:7080/${service}/articles/c0000000000o`}
-          id="c0000000000o"
-          isAmp={false}
-          pageType="article"
-          service={service}
-          statusCode={200}
-          pathname={`/${service}`}
-          variant={variant}
-        >
-          <ServiceContextProvider service={service} variant={variant}>
-            <MostReadContainer />
-          </ServiceContextProvider>
-        </RequestContextProvider>
-      </ToggleContext.Provider>,
-      container,
-    );
-  });
-
 describe('MostReadContainerCanonical', () => {
   beforeEach(() => {
     container = document.createElement('div');
@@ -83,6 +43,8 @@ describe('MostReadContainerCanonical', () => {
       fetch.mockResponse(JSON.stringify(mostReadData));
 
       await renderMostReadContainer({
+        container,
+        isAmp: false,
         service,
         variant,
         mostReadToggle: true,
@@ -97,6 +59,8 @@ describe('MostReadContainerCanonical', () => {
       fetch.mockResponse(JSON.stringify(setFreshPromoTimestamp(mostReadData)));
 
       await renderMostReadContainer({
+        container,
+        isAmp: false,
         service,
         variant,
         mostReadToggle: true,
@@ -119,6 +83,8 @@ describe('MostReadContainerCanonical', () => {
       fetch.mockResponse(JSON.stringify(setStalePromoTimestamp(mostReadData)));
 
       await renderMostReadContainer({
+        container,
+        isAmp: false,
         service,
         variant,
         mostReadToggle: true,
