+++ src/app/containers/MostRead/index.test.jsx
@@ -1,13 +1,9 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import { act } from 'react-dom/test-utils';
-import MostReadContainer from '.';
-import { RequestContextProvider } from '#contexts/RequestContext';
-import { ServiceContextProvider } from '#contexts/ServiceContext';
-import { ToggleContext } from '#contexts/ToggleContext';
 import newsMostReadData from '#data/news/mostRead';
 import zhongwenSimpMostReadData from '#data/zhongwen/mostRead/simp';
-import { setFreshPromoTimestamp } from './utilities/testHelper';
+import {
+  setFreshPromoTimestamp,
+  renderMostReadContainer,
+} from './utilities/testHelper';
 
 let container;
 
@@ -21,41 +17,6 @@ const services = {
     data: zhongwenSimpMostReadData,
   },
 };
-
-const getToggleState = enabled => ({
-  local: { mostRead: { enabled } },
-  test: { mostRead: { enabled } },
-});
-
-const renderMostReadContainer = async ({
-  isAmp,
-  service,
-  variant = null,
-  mostReadToggle = false,
-}) =>
-  act(async () => {
-    ReactDOM.render(
-      <ToggleContext.Provider
-        value={{ toggleState: getToggleState(mostReadToggle) }}
-      >
-        <RequestContextProvider
-          bbcOrigin={`http://localhost:7080/${service}/articles/c0000000000o`}
-          id="c0000000000o"
-          isAmp={isAmp}
-          pageType="article"
-          service={service}
-          statusCode={200}
-          pathname={`/${service}`}
-          variant={variant}
-        >
-          <ServiceContextProvider service={service} variant={variant}>
-            <MostReadContainer />
-          </ServiceContextProvider>
-        </RequestContextProvider>
-      </ToggleContext.Provider>,
-      container,
-    );
-  });
 
 describe('MostReadContainerCanonical', () => {
   beforeEach(() => {
@@ -75,6 +36,7 @@ describe('MostReadContainerCanonical', () => {
       fetch.mockResponse(JSON.stringify(setFreshPromoTimestamp(mostReadData)));
 
       await renderMostReadContainer({
+        container,
         isAmp: false,
         service,
         variant,
@@ -87,7 +49,12 @@ describe('MostReadContainerCanonical', () => {
     it(`should return empty string when mostRead toggle is disabled - ${service}`, async () => {
       const { variant, data: mostReadData } = services[service];
       fetch.mockResponse(JSON.stringify(mostReadData));
-      await renderMostReadContainer({ isAmp: false, service, variant });
+      await renderMostReadContainer({
+        container,
+        isAmp: false,
+        service,
+        variant,
+      });
       expect(container.innerHTML).toEqual('');
     });
   });
